reduce compilation times by breaking up swc_utils
